+++ src/configuration.js
@@ -1,5 +1,0 @@
-const vscode = require("vscode");
-
-module.exports = {
-    get: (property) => vscode.workspace.getConfiguration("justTesting").get(property)
-};