+++ src/helpers.js
@@ -1,0 +1,6 @@
+const vscode = require("vscode");
+
+module.exports = {
+    getSetting: (property) => vscode.workspace.getConfiguration("justTesting").get(property),
+    getActiveFile: () => vscode.workspace.asRelativePath(vscode.window.activeTextEditor.document.fileName)
+};