+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added "Run all tests" command
+- Added "Run all tests in file" command
 - Added "Run test on cursor" command
 
 ## 0.0.1